upgraded to sex-2.25, keeping sex-2.19.5 as fallback

+++ lib/compile_sextractor.sh
@@ -11,7 +11,9 @@ export MFLAGS=""
 
 VAST_DIR=$PWD
 TARGET_DIR=$VAST_DIR/lib
-LIBRARY_SOURCE=$VAST_DIR/src/sextractor-2.19.5
+#LIBRARY_SOURCE=$VAST_DIR/src/sextractor-2.19.5
+#LIBRARY_SOURCE=$VAST_DIR/src/sextractor-2.25.0_fix_disable_model_fitting
+LIBRARY_SOURCES="$VAST_DIR/src/sextractor-2.25.0_fix_disable_model_fitting $VAST_DIR/src/sextractor-2.19.5"
 
 #
 PATH_TO_THIS_SCRIPT=`readlink -f $0`
@@ -25,12 +27,23 @@ CFLAGS="-O2 -Wno-error $MARCH"
 echo " "
 echo -e "Starting script \033[01;32m$0 $1\033[00m"
 
+# Do nothing if there is a system-wide installation of SExtractor
+command -v sex &>/dev/null
+if [ $? -eq 0 ];then
+ echo "Found a system-wide installation of SExtractor, will do nothing"
+ exit
+fi
+
 
 if [ "$1" = "clean" ];then
  echo -e "\033[01;34mRemoving the local copy of SExtractor\033[00m"
- cd $LIBRARY_SOURCE
- make clean
- rm -f $TARGET_DIR/bin/sex
+ for LIBRARY_SOURCE in $LIBRARY_SOURCES ;do
+  cd $LIBRARY_SOURCE
+  make clean
+ done
+ if [ -f $TARGET_DIR/bin/sex ];then
+  rm -f $TARGET_DIR/bin/sex
+ fi
  echo "Script $0 is done."
  echo " "
  exit
@@ -39,40 +52,66 @@ fi
 echo -e "\033[01;34mCompiling the local copy of SExtractor\033[00m"
 echo "Using C compiler: $C_COMPILER"
 
-cd $LIBRARY_SOURCE
-make clean
-# --disable-model-fitting configure option turns-off model-fitting
-# features and allows compiling SExtractor without the ATLAS and FFTW libraries.
-./configure --disable-model-fitting --prefix=$TARGET_DIR CFLAGS="$CFLAGS"
-if [ ! $? ];then
- echo "VaST installation problem: an error occurred while configuring SExtractor
+# We maintain two versions of SExtractor
+for LIBRARY_SOURCE in $LIBRARY_SOURCES ;do
+ echo "######### Compiling $LIBRARY_SOURCE #########"
+ sleep 2
+ cd $LIBRARY_SOURCE
+ make clean
+ make distclean
+ # for sextractor-2.25.0
+ if [ -x ./autogen.sh ];then
+  ./autogen.sh
+  if [ $? -ne 0 ];then
+   echo "ERROR running autogen.sh"
+   continue
+  fi
+ fi
+ # --disable-model-fitting configure option turns-off model-fitting
+ # features and allows compiling SExtractor without the ATLAS and FFTW libraries.
+ ./configure --disable-model-fitting --prefix=$TARGET_DIR CFLAGS="$CFLAGS"
+ if [ $? -ne 0 ];then
+  echo "VaST installation problem: an error occurred while configuring SExtractor
 This should not have happened! Please report the problem (including the above error messages)
 to the VaST developer Kirill Sokolovsky <kirx@scan.sai.msu.ru>. 
 Thank you and sorry for the inconvenience."
- exit 1
-fi
-make -j9
-if [ ! $? ];then
- echo "VaST installation problem: an error occurred while compiling SExtractor
+  #exit 1
+  continue
+ fi
+ make -j9
+ if [ $? -ne 0 ];then
+  echo "VaST installation problem: an error occurred while compiling SExtractor
 This should not have happened! Please report the problem (including the above error messages)
 to the VaST developer Kirill Sokolovsky <kirx@scan.sai.msu.ru>. 
 Thank you and sorry for the inconvenience."
- exit 1
-fi
-make install
-if [ ! $? ];then
- echo "VaST installation problem: an error occurred while installing SExtractor
+  #exit 1
+  continue
+ fi
+ make install
+ if [ $? -ne 0 ];then
+  echo "VaST installation problem: an error occurred while installing SExtractor
 This should not have happened! Please report the problem (including the above error messages)
 to the VaST developer Kirill Sokolovsky <kirx@scan.sai.msu.ru>. 
 Thank you and sorry for the inconvenience."
- exit 1
-fi
+  #exit 1
+  continue
+ fi
 
-# Clean the source tree to save space
-make clean
-make distclean  
+ # Clean the source tree to save space
+ make clean
+ make distclean  
 
-cd $VAST_DIR
+ cd $VAST_DIR
+
+ 
+
+
+ # if we are here - assume the current SExtractor version compiled alright
+ break
+
+done # for LIBRARY_SOURCE in $LIBRARY_SOURCES ;do
+
+
 
 # Test if executable files were actually created?
 COMPILATION_ERROR=0
@@ -86,12 +125,14 @@ for TEST_FILE in $TARGET_DIR/bin/sex ;do
   echo -ne "\033[01;32mOk\033[00m,   "
  fi
 done
-echo "done!"
 
 if [ $COMPILATION_ERROR -eq 1 ];then
  echo -e "\033[01;31mCOMPILATION ERROR\033[00m"
  exit 1
 fi 
+
+echo "done!"
+
 
 echo -e "\033[01;34mFinished compiling SExtractor\033[00m"
 echo " "
